Successfully corroborated LFA and mean_energy models.
Working on getting mean energy model coupled with the liquid phase. If
we can demonstrate the same density dependence as a function of liquid
permittivty and liquid electron transport coefficients with the
mean_energy model as we did with the LFA approximation, than we're in
big frieking business. Let's get going on this.